Website/requirements.txt
WTForms==2.2.1
Flask==1.1.0
Flask_WTF==0.14.2
mysql_connector_repackaged==0.3.1
mysql-connector-python==8.0.17
adal==1.2.0
requests==2.21.0